requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.4.3
    # via aiohttp
aiohttp==3.11.8
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
attrs==24.2.0
    # via aiohttp
certifi==2024.8.30
    # via
    #   elastic-transport
    #   requests
charset-normalizer==3.4.0
    # via requests
datasets==3.1.0
    # via -r requirements.in
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
elastic-transport==8.15.1
    # via elasticsearch
elasticsearch==8.16.0
    # via -r requirements.in
filelock==3.16.1
    # via
    #   datasets
    #   huggingface-hub
frozenlist==1.5.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.9.0
    # via
    #   datasets
    #   huggingface-hub
huggingface-hub==0.26.3
    # via datasets
idna==3.10
    # via
    #   requests
    #   yarl
multidict==6.1.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
numpy==2.1.3
    # via
    #   datasets
    #   pandas
packaging==24.2
    # via
    #   datasets
    #   huggingface-hub
pandas==2.2.3
    # via datasets
propcache==0.2.0
    # via
    #   aiohttp
    #   yarl
pyarrow==18.1.0
    # via datasets
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via
    #   datasets
    #   huggingface-hub
requests==2.32.3
    # via
    #   datasets
    #   huggingface-hub
six==1.16.0
    # via python-dateutil
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
typing-extensions==4.12.2
    # via huggingface-hub
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via
    #   elastic-transport
    #   requests
xxhash==3.5.0
    # via datasets
yarl==1.18.0
    # via aiohttp
